refactor: MypagePage -> MyPage로 이름 변경
