+++ src/components/pages/index.tsx
@@ -5,6 +5,6 @@ export { default as Login } from "components/pages/Login/Login";
 export { default as SearchResultPage } from "components/pages/SearchResultPage/SearchResultPage";
 export { default as StoreDetailPage } from "components/pages/StoreDetailPage/StoreDetailPage";
 export { default as StoreDemandPage } from "components/pages/StoreDemandPage/StoreDemandPage";
-export { default as MypagePage } from "components/pages/MypagePage/MypagePage";
-export { default as BookmarkListPage } from "components/pages/MypagePage/BookmarkListPage/BookmarkListPage";
-export { default as MyReviewListPage } from "components/pages/MypagePage/MyReviewListPage/MyReviewListPage";
+export { default as MyPage } from "components/pages/MyPage/MyPage";
+export { default as BookmarkListPage } from "components/pages/MyPage/BookmarkListPage/BookmarkListPage";
+export { default as MyReviewListPage } from "components/pages/MyPage/MyReviewListPage/MyReviewListPage";
